% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figletr.R
\name{figlet}
\alias{figlet}
\title{Write a message with figlet fonts}
\usage{
figlet(msg, font = "standard")
}
\arguments{
\item{msg}{The message. An object which can be converted to character.}

\item{font}{A string specifying the font to use, if is not and option in the
available fonts (`figletr::fonts`) this parameter will be used as a path to
use this with external figlet fonts.}
}
\description{
Write a message with figlet fonts
}
