% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figletr.R
\name{figlet_demo}
\alias{figlet_demo}
\title{Test and show the available fonts in the package}
\usage{
figlet_demo(msg = "Figlet in R!")
}
\arguments{
\item{msg}{A string to test the fonts}
}
\description{
Test and show the available fonts in the package
}
