% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figletr.R
\name{font_info}
\alias{font_info}
\title{Parse the header of a figlet font}
\usage{
font_info(path = "inst/fonts/basic.flf")
}
\arguments{
\item{path}{The font path.}
}
\description{
Parse the header of a figlet font
}
