% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fonts}
\alias{fonts}
\title{Details of available font}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148 rows and 8 columns.
}
\source{
\url{http://www.figlet.org/fontdb.cgi}
}
\usage{
fonts
}
\description{
A dataset containing information about the fonts in the figletr package.
}
\keyword{datasets}
