% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figletr.R
\name{parse_font}
\alias{parse_font}
\title{Parse figlet fonts}
\source{
\url{http://www.jave.de/figlet/figfont.html}
}
\usage{
parse_font(path = "inst/fonts/basic.flf")
}
\arguments{
\item{path}{The font path.}
}
\value{
A named list which every elements is a character in the figlet font.
}
\description{
Parse figlet fonts
}
